requirements.txt
# Core dependencies
python>=3.9

# Google Earth Engine
earthengine-api==0.1.384
geemap==0.30.4

# Claude AI SDK
anthropic==0.18.1

# Geospatial processing
rasterio==1.3.9
geopandas==0.14.2
shapely==2.0.2
fiona==1.9.5

# Data processing
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
folium==0.15.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
requests==2.31.0

# Development tools
pytest==7.4.4
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
ipykernel==6.28.0
jupyter==1.0.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Logging and monitoring
loguru==0.7.2
sentry-sdk==1.39.2